sso: Make discoveryProtocol attribute removable
Setting a discoveryProtocol if a single IdP is used leads to the
following error:

"ERROR Shibboleth.Application : SSO discoveryProtocol specified without
discoveryURL"

Therefore the discoveryProtocol attribute has to be removable. Overriding
the default parameter $::shibboleth::params::discovery_protocol with undef
and applying a similar logic as for $discoveryURL and $idpURL does not
work. The reason is explained in

https://tickets.puppetlabs.com/browse/PUP-5295

"The semantics of giving undef as an argument to a resource (class or
other) is that the result is the default value of that parameter."

For this reason the value "none" was introduced for the
$discovery_protocol parameter to indicate removal of this attribute.